chore: remove circular dependencies
