+++ src/utils/user-agent-utils.ts
@@ -1,5 +1,5 @@
-import { includes } from './index'
 import { isFunction, isUndefined } from './type-utils'
+import { includes } from './string-utils'
 
 /**
  * this device detection code is (at time of writing) about 3% of the size of the entire library
